Remove include to file that does not exist and sign comparison compiler warning.
Review URL: http://codereview.chromium.org/6218
Patch from icefox.


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@3049 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/glue/password_form_dom_manager.cc
@@ -6,7 +6,6 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "config.h"
 
 #pragma warning(push, 0)
-#include "csshelper.h"
 #include "Document.h"
 #include "DocumentLoader.h"
 #include "Frame.h"
@@ -26,7 +25,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 // Maximum number of password fields we will observe before throwing our
 // hands in the air and giving up with a given form.
-static const int kMaxPasswords = 3;
+static const size_t kMaxPasswords = 3;
 
 PasswordForm* PasswordFormDomManager::CreatePasswordForm(
     WebCore::HTMLFormElement* form) {
